Changes to volunteer matching, navbar, and urgency renamed volunteerMatching to eventManagement for clarity. Updated navbar to reflect this. changed urgency from numeric values 4-1 (high to low) to critical >> high >> medium >> low for clarity. Added input validation for creating and editing events where all fields are required, event name is alphanurmeric with a set of allowed symbols, date field can select dates 3 days past current date and onwards, and the other fields are just required without restriction on input
