+++ frontend/src/app/events/page.jsx
@@ -10,7 +10,7 @@ const sampleEvents = [
       start: new Date(2025, 9, 29),
       end: new Date(2025, 9, 29),
     },
-    urgency: "4",
+    urgency: "Critical",
     description: "Help needed to direct traffic for the home game against Rice University. Needed from 5pm to 9pm for directing traffic in and out.",
     image: "/images/events/traffic-director.jpg",
   },
@@ -21,7 +21,7 @@ const sampleEvents = [
       start: new Date(2025, 10, 1),
       end: new Date(2025, 10, 1),
     },
-    urgency: "3",
+    urgency: "High",
     description: "Help needed cleaning the souls of the students who failed their OS exam in the agnes arnold auditorium because they didnt study semaphores enough",
     image: "/images/events/soul-cleanup.jpg",
   },
@@ -32,7 +32,7 @@ const sampleEvents = [
       start: new Date(2025, 10, 13),
       end: new Date(2025, 10, 15),
     },
-    urgency: "2",
+    urgency: "Medium",
     description: "PGH is outdated and needs to be rebuilt. Easy task should take 2 days",
     image: "/images/events/rebuild-pgh.jpg",
   },
@@ -43,7 +43,7 @@ const sampleEvents = [
       start: new Date(2025, 10, 16),
       end: new Date(2025, 10, 16),
     },
-    urgency: "2",
+    urgency: "Medium",
     description: "Looking for vendors to setup tents in the green between PGH and the library, must sell allowed goods such as food, clothing, or jewelry. Must be friendly towards students and report all revenue and provide the campus 25% of total profits",
     image: "/images/events/popup-shop.jpg",
   },
